Refactor routing into App.js
This should allow routes to work when hosted via Firebase
