+++ src/App.js
@@ -1,20 +1,24 @@
 import React from "react"
+import { Switch, Route, BrowserRouter as Router } from "react-router-dom"
 import winston from "winston"
-import { WinstonProvider } from "winston-react"
-import Home from "./components/pages/Home"
+import AppHome from "./components/pages/AppHome"
+import Mix from "./components/pages/Mix"
+import ColourDifference from "./components/pages/ColourDifference"
+import ColourMatch from "./components/pages/ColourMatch"
+import Footer from "./components/navigation/Footer"
 import "./App.css"
 
-const logger = winston.createLogger({
-	transports: [new winston.transports.Console()]
-})
-
 function App() {
-	return (
-		<WinstonProvider logger={logger}>
-			<div>
-				<Home />
-			</div>
-		</WinstonProvider>
+	return  (
+		<div>
+			<Router>
+				<Switch><Route exact path="/" component={AppHome} /></Switch>
+				<Switch><Route exact path="/Mix/:id" component={Mix} /></Switch>
+				<Switch><Route exact path="/ColourDifference" component={ColourDifference} /></Switch>
+				<Switch><Route exact path="/ColourMatch" component={ColourMatch} /></Switch>
+				<Footer/>
+			</Router>
+		</div>
 	)
 }
 export default App
